refactor(PullRequest(2)):
last Changes from reviewer 2
