+++ diversity-lunch-be/docker/.env.sample
@@ -13,3 +13,4 @@ CLIENT_ID=YourAzureClientId
 DIVERSITY_LUNCH_USER_ID="exampleUserId"
 CLIENT_SECRET="clientSecret"
 ADMIN_GROUP_NAME="adminGroup"
+BASE_URL="https://placeholder.com";
