+++ diversity-lunch-be/src/main/java/de/brockhausag/diversitylunchspringboot/profile/mapper/ProjectMapper.java
@@ -2,7 +2,7 @@ package de.brockhausag.diversitylunchspringboot.profile.mapper;
 
 import de.brockhausag.diversitylunchspringboot.profile.model.dtos.ProjectDto;
 import de.brockhausag.diversitylunchspringboot.profile.model.entities.ProjectEntity;
-import de.brockhausag.diversitylunchspringboot.profile.utils.Mapper;
+import de.brockhausag.diversitylunchspringboot.utils.mapper.Mapper;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
