+++ diversity-lunch-be/src/main/java/de/brockhausag/diversitylunchspringboot/profile/utils/baseApi/GenericBaseModelController.java
@@ -1,6 +1,6 @@
 package de.brockhausag.diversitylunchspringboot.profile.utils.baseApi;
 
-import de.brockhausag.diversitylunchspringboot.profile.utils.Mapper;
+import de.brockhausag.diversitylunchspringboot.utils.mapper.Mapper;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
