+++ diversity-lunch-be/src/test/java/de/brockhausag/diversitylunchspringboot/integrationstests/EMailControllerIT.java
@@ -84,7 +84,7 @@ public class EMailControllerIT {
 
         mockMvcSecurity = MockMvcBuilders.webAppContextSetup(appContext).apply(springSecurity()).build();
 
-        when(this.diversityLunchMailProperties.getSender()).thenReturn("diversitylunchtest@brockhaus-ag.de");
+        when(this.diversityLunchMailProperties.getSender()).thenReturn("test@example.com");
         when(microsoftGraphService.getGroups()).thenReturn(Optional.of(new ArrayList<>()));
         myProfileEntity = profileTestdataFactory.createNewMaxProfile();
         accountEntity = accountService.getOrCreateAccount(myProfileEntity.getEmail());
@@ -102,7 +102,7 @@ public class EMailControllerIT {
     @SneakyThrows
     @Test
     public void testSendTestMail_expectedToNotThrowException() {
-        when(this.diversityLunchMailProperties.getSender()).thenReturn("diversitylunchtest@brockhaus-ag.de");
+        when(this.diversityLunchMailProperties.getSender()).thenReturn("test@example.com");
 
 
         mockMvc.perform(
@@ -114,7 +114,7 @@ public class EMailControllerIT {
     @Test
     public void testSendTestMail_expectedCorrectTestContentInMail() {
         //Send email
-        when(this.diversityLunchMailProperties.getSender()).thenReturn("diversitylunchtest@brockhaus-ag.de");
+        when(this.diversityLunchMailProperties.getSender()).thenReturn("test@example.com");
 
         mockMvc.perform(
                 post("/api/mailing/sendTestMail")
@@ -143,7 +143,7 @@ public class EMailControllerIT {
 
         //Assert
         Assertions.assertEquals("test@test.de",to);
-        Assertions.assertEquals("diversitylunchtest@brockhaus-ag.de", from);
+        Assertions.assertEquals("test@example.com", from);
         Assertions.assertEquals("Testsubject",subject);
         Assertions.assertNotEquals("", body);
         Assertions.assertNotEquals(null, body);
