+++ diversity-lunch-be/src/test/java/de/brockhausag/diversitylunchspringboot/match/service/MatchingServiceTest.java
@@ -51,7 +51,6 @@ class MatchingServiceTest {
     @Mock
     DiversityLunchEMailService mockedEMailService;
 
-    DiversityLunchEMailService eMailService;
     @Mock
     MsTeamsService msTeamsService;
