+++ diversity-lunch-be/src/test/java/de/brockhausag/diversitylunchspringboot/profile/controllerTest/GenericBaseModelControllerTest.java
@@ -3,7 +3,7 @@ package de.brockhausag.diversitylunchspringboot.profile.controllerTest;
 import de.brockhausag.diversitylunchspringboot.dataFactories.BaseModelTestDataFactory;
 import de.brockhausag.diversitylunchspringboot.dataFactories.TestBaseDto;
 import de.brockhausag.diversitylunchspringboot.dataFactories.TestBaseEntity;
-import de.brockhausag.diversitylunchspringboot.profile.utils.Mapper;
+import de.brockhausag.diversitylunchspringboot.utils.mapper.Mapper;
 import de.brockhausag.diversitylunchspringboot.profile.utils.baseApi.GenericBaseModelController;
 import de.brockhausag.diversitylunchspringboot.profile.utils.baseApi.GenericBaseEntityService;
 import org.junit.jupiter.api.BeforeEach;
