+++ diversity-lunch-be/src/test/resources/application.yml
@@ -25,3 +25,5 @@ spring:
 diversity:
   ms-teams:
     time-zone: "Europe/Berlin"
+  url:
+    baseUrl: "https://example.com"
