+++ diversity-lunch-fe/src/components/AdminPanel/__test__/VoucherUpload.test.tsx
@@ -6,6 +6,7 @@ import React, { FC, useEffect } from 'react';
 
 import { BrowserRouter } from 'react-router-dom';
 import { Provider, useDispatch } from 'react-redux';
+import userEvent from '@testing-library/user-event';
 import * as fetcher from '../../../utils/fetch.utils';
 import { APP_STORE } from '../../../store/Store';
 import { loadAccount } from '../../../data/account/account.actions';
@@ -40,6 +41,11 @@ describe('VoucherUpload', () => {
         jest.spyOn(fetcher, 'authenticatedFetchPostCsv')
             .mockImplementation(async () => new Response('', { status: 200, statusText: 'OK' }));
         renderContainer();
+        const fakeFile = new File(['hello'], 'hello.csv', { type: '.csv' });
+        const fileInput = await screen.findByPlaceholderText('Datei hier reinziehen');
+        act(() => {
+            userEvent.upload(fileInput, fakeFile);
+        });
         const accessCodeButton = await screen.findByText('Upload');
         act(() => {
             fireEvent.click(accessCodeButton);
